+++ packages/graph-explorer/src/components/Tabular/ModuleContainerTabularHeader/index.ts
@@ -1,2 +1,0 @@
-export { default as ModuleContainerTabularHeader } from "./ModuleContainerTabularHeader";
-export type { EntitiesTabularHeaderProps } from "./ModuleContainerTabularHeader";