+++ packages/graph-explorer/src/modules/EntitiesTabular/EntitiesTabular.tsx
@@ -1,17 +1,31 @@
 import { useState } from "react";
-import { Panel } from "@/components";
+import {
+  GridIcon,
+  Panel,
+  PanelHeader,
+  PanelHeaderActions,
+  PanelHeaderCloseButton,
+  PanelHeaderDivider,
+  PanelTitle,
+  Select,
+  SelectContent,
+  SelectItem,
+  SelectTrigger,
+  SelectValue,
+} from "@/components";
 import type { TabularInstance } from "@/components/Tabular";
-import { ModuleContainerTabularHeader } from "@/components/Tabular";
+import { ExportControl } from "@/components/Tabular";
 import TabularControlsProvider from "@/components/Tabular/TabularControlsProvider";
 import useTranslations from "@/hooks/useTranslations";
 import { EdgesTabular, NodesTabular } from "./components";
+import { useViewToggles } from "@/core";
 
 enum TableId {
   edges = "edges",
   nodes = "nodes",
 }
 
-const EntitiesTabular = () => {
+function EntitiesTabular() {
   const t = useTranslations();
 
   const [selectedTable, setSelectedTable] = useState<TableId>(TableId.nodes);
@@ -52,22 +66,37 @@ const EntitiesTabular = () => {
     <Panel>
       {nodeInstance && edgeInstance && selectedTabularInstance && (
         <TabularControlsProvider tabularInstance={selectedTabularInstance}>
-          <ModuleContainerTabularHeader
-            tables={[
-              {
-                value: TableId.nodes,
-                label: t("entities-tabular.all-nodes"),
-              },
-              {
-                value: TableId.edges,
-                label: t("entities-tabular.all-edges"),
-              },
-            ]}
-            selectedTable={selectedTable}
-            onTableChange={tableId => {
-              setSelectedTable(tableId as TableId);
-            }}
-          />
+          <PanelHeader>
+            <PanelTitle>
+              <GridIcon className="icon" />
+              Table View
+            </PanelTitle>
+            <PanelHeaderActions>
+              <Select
+                aria-label="Table"
+                value={selectedTable}
+                onValueChange={tableId => {
+                  setSelectedTable(tableId as TableId);
+                }}
+              >
+                <SelectTrigger className="w-36">
+                  <SelectValue placeholder="Select an entity type" />
+                </SelectTrigger>
+                <SelectContent>
+                  <SelectItem value={TableId.nodes}>
+                    {t("entities-tabular.all-nodes")}
+                  </SelectItem>
+                  <SelectItem value={TableId.edges}>
+                    {t("entities-tabular.all-edges")}
+                  </SelectItem>
+                </SelectContent>
+              </Select>
+              <div className="grow" />
+              <ExportControl />
+              <PanelHeaderDivider />
+              <CloseButton />
+            </PanelHeaderActions>
+          </PanelHeader>
         </TabularControlsProvider>
       )}
       {tableList.map(table => (
@@ -85,6 +114,12 @@ const EntitiesTabular = () => {
       ))}
     </Panel>
   );
-};
+}
+
+function CloseButton() {
+  const { toggleTableVisibility } = useViewToggles();
+
+  return <PanelHeaderCloseButton onClose={toggleTableVisibility} />;
+}
 
 export default EntitiesTabular;