+++ Changelog.md
@@ -5,6 +5,8 @@
 - **Added** support for openCypher in the query editor
   ([#1016](https://github.com/aws/graph-explorer/pull/1016),
   [#1024](https://github.com/aws/graph-explorer/pull/1024))
+- **Added** close button to table view
+  ([#1026](https://github.com/aws/graph-explorer/pull/1026))
 - **Updated** dependencies to latest versions and some minor refactorings
   ([#1014](https://github.com/aws/graph-explorer/pull/1014),
   [#1023](https://github.com/aws/graph-explorer/pull/1023),